+++ src/components/Home.js
@@ -21,7 +21,6 @@ const Home = () => {
     event.preventDefault();
     setImage(event.target.files[0]);
   };
-  console.log(image);
   return (
     <div>
       <h1>Welcome to instaHarvest</h1>
