+++ src/components/product/AddProduct.css
@@ -20,11 +20,11 @@
     box-sizing: border-box;
     display: block;
   }
-  .add-product button {
+  /* .add-product button {
     background: orangered;
     color: #fff;
     border: 0;
     padding: 8px;
     border-radius: 8px;
     cursor: pointer;
-  }
+  } */
