+++ src/components/product/UserProducts.js
@@ -1,49 +1,51 @@
 import { useRequest } from "../../hooks/hooks";
-import { useEffect, useState } from "react"; 
-import Spinner from '../UI/Spinner';
-import Product from './Product'; 
-
+import { useEffect, useState } from "react";
+import Spinner from "../UI/Spinner";
+import Product from "./Product";
 
 const UserProducts = () => {
+  const [userProducts, setUserProducts] = useState([]);
+  const [isLoading, data, error, errorNum, sendRequest] = useRequest();
 
-    const [userProducts, setUserProducts] = useState([]);
-    const [isLoading, data, error, errorNum, sendRequest] = useRequest(); 
+  const getProducts = () => {
+    sendRequest("/api/products/products-per-user", "post", null, true);
+  };
 
-    const getProducts = () => {
-        sendRequest("/api/products/products-per-user", "post", null, true)
+  useEffect(() => {
+    getProducts();
+  }, []);
+
+  useEffect(() => {
+    if (data && data.msg === "deleted") {
+      getProducts();
+    } else if (data) {
+      setUserProducts(data.user_products);
     }
+  }, [data]);
 
-    useEffect(() => {
-        getProducts();
-    }, []); 
+  const handleDelete = (product_id) => {
+    sendRequest(
+      "/api/products/delete_product",
+      "delete",
+      { product_id: product_id },
+      true
+    );
+  };
 
-    useEffect(() => {
-        if (data && data.msg === "deleted"){
-            getProducts();
-        }
-        else if(data) {
-            setUserProducts(data.user_products)
-        }
-    }, [data])
+  return (
+    <div>
+      <h2>All your products are here! </h2>
+      {isLoading && <Spinner />}
+      {userProducts &&
+        userProducts.map((product) => {
+          return (
+            <div key={product.product_id}>
+              <Product product={product} onDelete={handleDelete} />
+            </div>
+          );
+        })}
+    </div>
+  );
+};
 
-    const handleDelete = (product_id) => {
-            sendRequest("/api/products/delete_product", "delete", {product_id: product_id}, true)
-    }
-
-    console.log(userProducts)
-    return(
-        <div>
-            <h2>All your products are here! </h2>
-            {isLoading && <Spinner />}
-            {userProducts && userProducts.map((product) => {
-                return(
-                    <div key={product.product_id}>
-                        <Product product={product} onDelete={handleDelete}/>
-                    </div>
-                )
-            })}
-        </div>
-    )
-}
-
-export default UserProducts; 
+export default UserProducts;
