+++ src/components/profile/ProfileEditTab.js
@@ -1,0 +1,3 @@
+const ProfileEditTab = () => {};
+
+export default ProfileEditTab;
