--- src/components/profile/ProfileTab.js
@@ -1,3 +1,0 @@
-const ProfileTab = () => {};
-
-export default ProfileTab;
